+++ src/language/routing/index.ts
@@ -1,5 +1,8 @@
 export { createLocaleBootScript } from "./boot.js";
+export { createLocaleDocumentBody } from "./document.js";
+export type { LocaleDocumentBodyOptions } from "./document.js";
 export {
+  DEFAULT_LOCALE_COOKIE_NAME,
   DEFAULT_LOCALE_STORAGE_KEY,
   cleanLocale,
   matchLocale,
@@ -8,20 +11,13 @@ export {
 } from "./options.js";
 export type { LocaleRouting, LocaleRoutingOptions } from "./options.js";
 export {
-  buildLocalePathname,
-  localeShellOutFile,
-  localeShellRoutes,
-  normalizePathname,
-  parseLocalePathname,
-} from "./path.js";
-export type { LocaleShellRoute, ParsedLocalePathname } from "./path.js";
-export {
   configureLocaleRouting,
   currentLocale,
   getLocaleRouting,
-  localeHref,
   onLocaleChanged,
   persistLocale,
   setCurrentLocale,
 } from "./runtime.js";
 export type { LocaleListener } from "./runtime.js";
+export { applyLocaleMeta, applyLocaleView, readLocaleMeta } from "./view.js";
+export type { LocaleMeta } from "./view.js";